# release-23.1: colexecerror: avoid debug.Stack in CatchVectorizedRuntimeError